Add blueprint update script.
This script can be run from a gerrit hook to update Launchpad
blueprints related to a given change.  It looks for items like
"bp foo", "blueprint foo", "bp: foo", "blueprint: foo", etc,
and updates the foo blueprint of the current project to add a
whiteboard note that links back to the change.

It will also similarly update a blueprint if the topic of the
change (after any initial text up to the first '/') is the name
of a blueprint.

This has a corresponding puppet configuration change to put it
into production.

Change-Id: I62f965ca7f14f589e3b299ea46729efb68abd06f
Reviewed-on: https://review.openstack.org/253
Reviewed-by: Monty Taylor <mordred@inaugust.com>
Tested-by: Jenkins