Use a new image name each time we snapshot.
Work around a bug in RS cloud servers where after deleting an image,
you may not be able to create a new one with the same name.  Instead,
generate a unique name for every image, and when we go to use it, find
an image that _starts_ with the name we're looking for.

Change-Id: I2e674c544b96e976c0460298ab74de1da8e7144c