Cache a copy of all images on the template VM.
This is part 1 of a 2 part change that will get the images
pre-populated in the devstack directory on the vm so that
they aren't downloaded via the Internet each time.

This needs to be merged first so that the directory with the
images is available on the vm.  The next change, to the
gate-host script, will expect it to be there.

Change-Id: I230e3d3032a9d605b2e1192e1f1df6d5da75d01e