Several improvements to integration testing.
Change libcloud imports to remove deprecated module paths.

Actually move the pip cache into the place devstack expects on
the devstack host.

As soon as the new version of openstack-ci is checked out, start
using it for the scripts in the currently running test, so that
we are testing as much of the test infrastructure itself as
possible.

Allow the update and launch scripts to have the server/image name
specified in an environment variable for easier
testing/upgrading.

Change-Id: Iee0a946afa80929e99c08e5a6b0504c1d8dd0093