Document changes to forge permissions.
The main problems with restricting forging of author or committer
are around cherry-picking and submitting patches on behalf of third
parties.

When cherry-picking, normally the Author of a commit will be kept,
and the committer changed.  That means that to support cherry-picking,
we need to allow anyone to forge the author identity, but not the
committer id.

This change permits that.

Note that all contributors are required to sign the CLA.  If we allow
forging of author identities, we open a hole where a person who has
signed the CLA could submit a patch authored by someone who has not.

However, in general people are expected to upload their own changes
(except for cherry-picking across branches, and those changes have
already been uploaded by a person who signed the CLA).  So in practice
we wouldn't expect a change submitted on behalf of a third party.
If we want to support easily cherry-picking, we'll have to accept that
and inform developers of the behavioral expectation.

At least by not allowing the forging of committer identities, there is
still a person who signed the CLA who is "on the hook" for that change.

Change-Id: I3893fd75d3dc3f724d5aae036b2108ce75409af8