Use bundles instead of relocatable virtualenvs.
This is step one. We'll create a venv the old way and also additionally a
bundle. If the bundle exists on the copy-venv side of things, we'll install
using it. Otherwise, we'll just do the old relocation code. After a day
or two (once all of the venv artifacts have bundles) we can remove the use
of install_venv and just go to pure bundles.

Change-Id: Ie56e270ac4944a29253609730b013166cae05c26