Abandon devstack vms that launch slowly.
If a vm takes more than 900 seconds to launch, abandon it (set it
to error state and let the reaper come along and delet it in 24 hours
when things have hopefully settled down).  Addresses bug 921738.

Set the number of vms in the pool to 10 (increase from 5) to paper
over more spurrious operational errors from cloud providers.

Change-Id: Ib000e420377f4d19871bd42fb360016a519e4b2f