Ensure that linux-headers-virtual is installed.
Also linux-headers-`uname -r`.  This is needed for the
stable/diablo branch.  And since apt-get doesn't know how to pull
down the running version, dkms builds can fail if this isn't
installed for the specific running version beforehand.  The
unversioned package will install the lastest version, but that
may not be the kernel the image is currently running.

Change-Id: Idb619ed57f6e101001fea3825c328b471d97e092