Change the label of devstack nodes on use.
Instead of disconnecting then node, change the label to
"devstack-used", which is a label we will never use for a job.
That way the node will not be re-used, but we don't get a warning
in the job console output about the node going offline.

Also, make sure the nodes are created for exclusive use
(tied jobs only).

Change-Id: Ib162d78baee17452cec13c150235f3584737d3c3