Ignore symbolic references during local_prep.
In the image update script ignore any branch returned by `git branch -a`
containing the ' -> 'substring. This entire string denotes a symbolic
reference and is not a valid branch (so it cannot be checked out). We
can do this safely because `git branch -a` will also return the branch
referred to by the reference.

Change-Id: I47852ffd6e25fe3a29a1b310142c5614ca5a6fa9