Compress logs before Jenkins can copy them.
Compress the various logs devstack gate collects before Jenkins can copy
them. This reduces the headroom necessary for all of the files on the
logs server by a significant amount. Also, to save even more space on
the logs server stop copying the screen logs as syslog appears to
capture all of this data.

Change-Id: I78f5bfc436c288d52aac2083ba8c24cff4a67648