Add option to ignore ZUUL_REF when setting up workspace.
When setting up the old workspace for grenade tests, we would
still check for the ZUUL_REF and therefore might end up checking
out a proposed change into the old repo instead of just leaving
it at stable/foo.

This is a new parameter to setup_workspace.  Currently it's set
to ignore zuul refs when setting up the old workspace, and use
them when setting up the new.  If we start doing forward grenade
testing (ie, a change to stable/folsom must not break upgrading
from stable/folsom to master or stable/grizzly), we may need
the option of reversing this.  But that's a ways away.

Move the grenade branch selection logic into devstack-vm-gate-wrap
so that it is easy to update as we progress along releases.  Make
DEVSTACK_GATE_GRENADE a binary variable like the other
DEVSTACK_GATE variables.

Also, fix renaming the log files to .txt. and removing duplicates
when we're an extra level deep in the log dir when running grenade.

Change-Id: Iee9530cf2a5482bf053ba72cd0448be5de0e2c1f