Build VM images when min_ready >= 0.
* devstack-vm-update-image.py: Previously we would not build new VM
images when the base_image.min_ready value was <= 0. Change this check
to only skip image updates when the min_ready value is < 0. This allows
us to have up to date images even when we do not actually want to have
any slave VMs running on that provider. You can still prevent image
building by setting min_ready to -1.

Change-Id: If63c2cffeca13f0f7281fe83432926d48323aa69