Document how to manually simulate infrastructure.
Fixes bug 1172428.

* README.rst: Developers often have a need to recreate gating
integration tests manually, and this provides a walkthrough of
making a DG-slave-like throwaway server without the overhead of
building other CI infrastructure to manage a pool of them. This can
be useful to reproduce and troubleshoot failures or tease out
nondeterministic bugs. The driving use case is for pre-testing
embargoed vulnerability fixes, increasing the chance they can get
through gating on the first try.

Change-Id: I0f78011c183f64b894b336eb8e0d25dc5ba00bde