override tempest http_image variable in the gate
one of the tempest tests is to ensure glance can download something
over http, defaulting to the upstream cirros image. In the gate
that url ends up unstable on the network.

To get around this, feed devstack a tempest variable which sets
this to the horizon url. That's a valid http url. Because we don't
test for content, this is good enough for the test.

Rev 2: hardcode horizon to 127.0.0.1, the previous variable wasn't
set.

Partial fix for bug #1190623 once the devstack change lands.

Change-Id: Ifd94ace4b07a5ab9d20af7088e38420e9f3aab88