Fix tempest coverage runs
Commit 8fe6db0a changed the report generation for coverage to use
the tox coverage job. However, this job wouldn't work for the devstack-gate.
We need to pass the output directory to tempest_coverage.py so
that the coverage report gets put in the logs directory. After posargs was
added to the tox job the needed arguments can be passed to the report command.
This commit adds the needed arguments to the coverage job and reworks the if
logic to correct the behavior for the coverage flag.

Change-Id: I93bb9d91637c262e4ece6580964541c40c96eb60