Increasing th backing volume size to 24GiB for parallel tests
The tempest parallel run will cause, more volume allocation at the same time.

The cinder lvm driver uses sparse file for storing the volumes,
so the 24GiB will not be really used, and the allocation time is almost
the same as for 10GiB.

The gate vm normally has enough free space for 24GiB loopback
volume files.

Change-Id: I287536c731f941293bf31d0befca3a65f8fa609f