Add testr run and use subunit2html with testr
This commits adds a new run option DEVSTACK_GATE_TEMPEST_TESTR_FULL
which is used to run the same tests as DEVSTACK_GATE_TEMPEST_FULL
but using testr --parallel instead of nose. This is going to eventually
be removed when using testr with tempest gets stable enough to gate
projects. But for now it will be a separate option. This also adds
support for running subunit2html on all testr runs to get pretty html
output from the testr runs.

Change-Id: I60e3ca48e8f4196503b09711997bd70e641bd16c