Add jeepyb to the list of cached repos
The pbr integration test turns out to also be a good test of the
contents of the requirements repo and of jeepyb's run-mirror script,
since it uses both. If we add jeepyb to the list of repos that are
on a devstack-gate node, then we can construct the integration test
to actually gate not only changes to pbr, but also changes to jeepyb
and requirements, which should ensure that we don't wind up with
bad requirements added or mirror breakages.

Change-Id: I8cea5f4df7563dcc7d44a046e2d3fb09e0d8fc9e