Handle Jenkins node already being added to Jenkins
If the devstack launch jobs timeout it is possible for some Jenkins
nodes to be in Jenkins but still marked as Building in the devstack gate
DB. Previously when this happened it caused the node to be deleted
possibly breaking any tests running on that node. Now if the node is
already in Jenkins ignore the error and allow the node to go into a
ready state.

Change-Id: Ib285f90f86bf79bab274b4e1a289284b0e5f30f9