add oslo.* to the projects list
We want to add oslo.* into the gating process, so git versions of
the libraries can be tested against the rest of OpenStack, and
that those parts of oslo can be tested against the gate. This is
a prereq of devstack changes we want to land to test fully against
global requirements.

Change-Id: Ib9abf6c9207c27d1e74726196f1bd19712901db7