Skip devstack/exercises by default
In order to avoid maintaining test cases in two location,
the devstack/exercises should be disabled when tempest is enabled.

The exercises is preserved for grande and stable branch jobs.

A smaller/faster 'smoke' test case set can be selected from the tempest
test cases if it is needed.

Fixing bug #1023131

Change-Id: Icac3998fb2c18274d909afcff8708af149d68197