Add support for large_ops tempest test
This test launches a bunch of instances at once (in attempt to find any
performance/scale issues).  This is the regression test for bug 1199433

Set TEMPEST_LARGE_OPS_NUMBER, the number of instances to attempt to spawn at
once to 150.  The number 150 is picked via trial and error.  Before the
related bug was fixed, one couldn't even spawn 50 instances.

Change-Id: Ic9b4872ae9ecccff8b4876b4c9d9572cb1233081