Revert "Enable q-vpn service"
This reverts commit abebbd61eb4758ca5cbab35f3494e7f0262e5bf9.

From the discussion in bug #1230407, the neutron deadlocks may be
a persistent latent issue with driving large amounts of rpc traffic
in neutron. The reason this became an issue this week may be related
to turning on q-vpn in the gate, as that generates even more traffic.

If we revert this change, and things go back to "normal" failure
rates, then this gives us more evidence in that direction.

Change-Id: Ia62dc1cd7efce3aaab5b8825e26082d42f629f88
Related-Bug: 1230407