Improve fallback to master branch
If a change to a stable branch follows a change to the master branch
but a project does not have a stable branch, the workspace setup
falls back to checking out the master branch.  However, it checks
out the actual current master branch rather than the proposed future
state of the master branch as determined by Zuul.  This change will
cause d-g to look for a Zuul ref for the master branch if it decides
to fall back.

Change-Id: Idadb0e0c4cd9cb120ec2c55f16282fc66076a61d