Call a script to check for ERRORs in logs on tempest success
Since this script does not exist pre-havana the call is skipped for those
test runs. Stress tests already do their own checks. The check_logs script
has a whitelist for existing errors that is maintained in tempest and
should be reduced as bugs causing spurious errors are fixed.

Partially implements blueprint fail-gate-on-log-errors

Change-Id: I21438ce6aa178c3e02ac98d98df9ed88a646d8ce