don't UNDO_REQUIREMENTS after we update from g-r
it turns out there was a loophole in the global-requirements to
rule them all approach, because tempest runs under tox, and the
default devstack behavior is to reset the source trees after
global-requirements was used for setup.py develop, so that we
wouldn't be poluting people's working trees.

This meant that if tempest contained an incompatible requirement
from g-r, it could wedge the gate (which it did for stable/havana).

We've since introduced UNDO_REQUIREMENTS in devstack that lets us
control this reset. This change in the gate will ensure that we
don't reset tempest back to a state where it can wedge us.

Change-Id: I6fb14cef9d52a35569c2f0215267609cd6741337