Let us gate grenade forward
This sets up all the grenade variables both needed to cover Havana
and Icehouse runs, but also to provide a new DEVSTACK_GATE_GRENADE_FORWARD
that let's us do a forward upgrade on a stable branch.

Also, get completely explicit on all the branches, to ensure that we
know what directions we are headed in.

Update setup_workspace to take local variables of branch name to try
to make the code a little clearer, and move all the setup_workspace to
the same point in the script.

Co-Authored-By: Mauro S. M. Rodrigues <maurosr@linux.vnet.ibm.com>
Change-Id: Iffcbd127c20780c5de6a16c41d1f0fd9f0d4d470