allow configdrive vs. metadata server to be configured
this gives us control of the config drive settings in devstack so
that we can force it off, to ensure that we are using the
metadata server instead for configuring the guests.

Needed because neutron does some extra clever things with the
metadata server setup which got broken because we were defaulting
to config drive in all the runs.

Depends-On: Ic2acb568b55a05b84120b181383ea45ea190dcad

Change-Id: Ib203bb0c812939f32b36731464b18566a187c85a