Fix copying of select-mirror outputs.
The Jenkins user was attempting to write to ~stack and failing because
it does not have permission to do so. Use more sudo to get around that.
Additionally move the fix for /etc/hosts into setup_host instead of
setup_workspace. This will make the use of sudo quieter as it will be
able to resolve $HOSTNAME earlier.

Change-Id: I8e6bd821ceac9d0614c068dc7ce7a6657329d84c