Drop large-ops test down to 50 instances from 100
It turns out the variance in cloud instances is very high, especially
when comparing different cloud providers and regions. This test was
originally added as a regression test for the nova-network issues with
rootwrap. At which time this test wouldn't pass for 30 instances.  So 50
is still a valid regression test.

Change-Id: I20a158cdd043d53be0b96437826559a71f7698bf