Set concurrency to num CPUs / 2
A recent change added a knob for setting the concurrency for tempest
runs.  It started out at 4 and then was reduced to 2.  Both of these
values assumed that all of our test nodes had 4 CPUs.  That's not
actually the case.  We now have some nodes with 8.

Change the default from 2 to the number of CPUs / 2.  This won't
change anything on most of our test nodes, but it will allow us to use
more on some nodes.

Change-Id: I559359b44dcbb717de3eab31cb296f033951f312