Copy tempest.conf file with log files
This commit adds the tempest conf file to the set of artifacts that get
copied after the tests are finished running. Previously the entirety
of the conf file was cat into the console output by devstack which is
difficult to parse. By copying it that cat tempest.conf can be removed
from devstack which will make it easier to verify the tempest conf
file from a test run.

Change-Id: Ifb36918cd0d686cb3865f5322cd62c209acaaf30