Run stack.sh with forced line buffering
Sometimes when stack.sh has an error we don't get all of the console log.
This is part of a set of patches to attempt to eliminate console output
buffering so Jenkins has an opportunity to get all of the output.
The other change is in https://review.openstack.org/#/c/71996/ that
eliminates the buffering that tee(1) does.

If this works and is not too harmful to run times we'll come back and
to the same thing to the grenade.sh execution.

Change-Id: I7f567c43c966fcfd822715ec0baa53755af6099a