start first partition at 1M not 0
trying to create a partition with parted at actually 0 offset
causes a warning on alignment. On a real disk this would be an
actual issue with offsets, and could have a real performance
impact. With a virtual disk, it's a little less clear, but can't
hurt to shave off 1M of swap to align in a more typical way.

Change-Id: I28cb60c635280603488a9b34ead848ee76ca111c