Allow configurable large-ops tests
This makes the large-ops job use the non-zero enable flag as the
number of resources to configure instead of hard-coding it into
the job. Since we can't merge this and the change to the job
at the same time, this includes special support for the =1 case to
run the same number we currently do (50). After both merge, the
extra condition can be cleaned up with a test for ">0".

Change-Id: I08c392bb620b5e911c721c806310749eeaba8ed0