turn back on log enforcement
This turns back on the log enforcement that certain logs need to
be clean. This will fail if a new log gets added that has errors.

The logic was cleaned up a bit to rely on -o errexit. While we
should already be at -o errexit at that part of the script, an
added call is made to ensure if anyone changes the top half they
don't change the behavior of this part.

We also create a specific out for 3rd party CI, or other tests
that don't want this enforcing, which is to set
DEVSTACK_GATE_CLEAN_LOGS=0 in the environment.

Change-Id: I50944ad64bec948504a0124ea423a0f96f8a8361