usability improvements to test-functions.sh
instead of dying at the first error found, run all the tests
and show all the errors. This makes it simpler to debug as you
can address more than one issue at once.

add the use of the bash 'caller' function to automatically print
out the name of the function and line number on both successful
and failing asserts. This ensures it's easy to get to the issue
at hand.

Change-Id: If15f50c6c41aa3940682d41c776589407a42bf4f