Put cached files in old/ dirs too
Previously because files in the cache were moved and the new/ workspaces
were configured first none of the cached files ended up in old/
workspaces. Make it possible to copy the cache the first time around so
that the second time through we can do the mv and get all of the files.

Change-Id: I4f41e3ebbde85a27515c5bd27f6a4c706ffe9bb4