Support for alternate trunk name (not master)
Commit 537a2926745a2acbacfa682295f06fa44c42716a introduced support
for having trunk named something other than master.

The name "master" is still hardcoded in the test-matrix, and in
the list of allowed branches. Adding support for an arbitrary
features file via DEVSTACK_GATE_FEATURE_MATRIX env variable.

The list of allowed branches is now parsed from the feature matrix.

Change-Id: Ic07c958e1630f7e62d9ff35b0add40a5d2b1548b