Capture old tempest artifacts too
During a grenade run we run tempest twice. However we were only
capturing the testr results and tempest.log of the second run. This
makes it slightly more difficult to debug a failure that occurred
during the first grenade tempest run. This commit address that by
capturing both sets are artifacts.

Change-Id: I6610251246933139c5ab63df5164acc1b4c069f9