Make one copy of grenade log files
Change 709f2349 removed the blanked 'rm' of grenade log files, but
that leaves behind several confusing entries in the final log output;
e.g [1].

---
 grenade.sh.log.2014-06-16-014311
 grenade.sh.log.2014-06-16-014311.2014-06-16-014311.summary
 grenade.sh.log.summary
 grenade.sh.txt
---

grenade.sh.txt is greande.sh.log renamed, which was a symlink to the
time-stampped file, which is no longer removed.  The summary files are
also duplicated and not directly visible because they do not have a
.log extension.

This change just keeps the grenade log and summary files.  It renames
the summary so it can be viewed, and removes the unnecessary files.
It also puts grenade's localrc in the top-level directory with the
logs, renamed to grenade_localrc

[1] http://logs.openstack.org/83/98283/2/gate/gate-grenade-dsvm/8ba1473/logs/

Change-Id: I932ee1583f04756579abf63c792d336ae937f64d