Fix typo in DEVSTACK_GATE_TEMPEST_NOVA_V3_API
This commit fixes a small oversight in setting the nova v3 api tempest
tests. When this env variable was added to set whether the testing was
enabled the check used a different variable name than the wrap script.
Also neither of these variables matched what was being set by jenkins,
this meant that the v3 api tests would never be enabled. This fixes
that by ensuring the variables are the same and matches the variable
set in the JJB template for the v3 api jobs.

Change-Id: I54925a468c4dc7eb9390a78d509b9385caa587d4