add network_sanity_check
a couple of quick early tests of the network sanity to try to help
us categorize the network quality from various zones at various
times of the day. As connectivity to pypi.openstack.org is one of
our biggest issues, use that as the test to see if this is useful.

Also test mirror.rackspace.com. Even though it's not always the
apt mirror we use, it does fail a bunch, and would be interesting
to know what it's base connectivity looks like when it does.

Change-Id: I71e4e732c8222e7cfe4775c32038eddd65f40b1b