Sanity check in setup_workspace
setup_workspace called with unquoted variables,
which could lead to wrong argument moved into wrong variable,
when one of the variables was empty string or contained space.

This change quotas the parameters for setup_workspace,
and also validates the
OVERRIDE_ZUUL_BRANCH, GRENADE_OLD_BRANCH, GRENADE_NEW_BRANCH are
not empty strings when they are used.

Change-Id: I13d1458c9cebd83b0ca700b3153f6645b2931c0d