Fix ssh key specification in the README.rst
The ssh public key entries starts with key type
specifier like 'ssh-rsa', but the puppet module which consuming the ssh_key
 expects only the key part.

Also switching to the id_rsa.pub in the example, because usually it
contains just only one public key, what the users usually pushes to the servers.

The authorized_keys file is for the keys accepted by the user's machine
itself.

Change-Id: Ic92729b572582360495b3544e479de98cd3c02c0