turn off clean log checking
The effort to clean out the issues in the logs in OpenStack has
largely stalled, and the impact of this is currently just a lot more
spurious failures that aren't getting looked at. Also it's kind of a
layer violation given that the checker is in Tempest but called from
devstack-gate.

The number of logs that were actually considered clean is actually
very small (only 4 or 5 out of 30), so this wasn't really preventing
many new bugs from entering. Removing it will decrease failure rates
for tempest runs which otherwise passed.

Change-Id: Id236826a1f2669c92b20cf8f07ee604f2c12f06c