Remove TESTR_FULL option for running tempest
The TESTR_FULL option was added back when tempest was running by
default with nosestests serially to test using testr in parallel.
However, this migration happened >1yr ago and the testr-full job is
just a duplicate of the regular full now. This commit removes the
option of running with TESTR_FULL to prevent confusion.

Change-Id: I57d1b249ced3d68143da7f514136d3b2e5deffd2