Set ZAQAR_BACKEND in localrc if present
In order to add more gates on different backends for Zaqar, we need to
set the ZAQAR_BACKEND env var in devstack. Setting it in the job is not
enough since we need to teach devstack-gate.sh what to do with it. This
patch does exactly that. The ZAQAR_BACKEND variable will be appended to
localrc if it's been set.

Change-Id: Ie629d6233798bd44663dfddb3ec02e3cdc1a69fe