Havana branches are dead, remove havana from d-g
We EOLed, tagged, then deleted the stable/havana branches everywhere.
Clean havana out of devstack-gate as this config is no longer necessary.

Note this adds Juno specific test-features.sh tests because the havana
tests are no longer needed but we should continue to test that we
generate the proper feature lists for our stable branches.

Change-Id: Iea93ac03a9729cf6fe75a6410de4dec2c531f8d2