Introduce OVERRIDE_ENABLED_SERVICES for func tests
We agreed at the Kilo summit that projects will write in-tree
functional tests, which may be executed in a targeted devstack
environment.
Introduce OVERRIDE_ENABLED_SERVICES so that projects which
want to used devstack for their test, may also reuse the d-g
scripts and fully customize the list of services installed
in their test clouds.

Change-Id: I4170166d9d0eff5995b330c85e0ce30592b0c0f4