Temporarily disable testing lbaas, fwaas, and vpnaas
As part of the migration of the advanced services out of the neutron repo
the testing in tempest needs to be disabled, otherwise the removal patch
in neutron will not be able to land. This is just a temporary workaround
to enable forward progress on the repo split, and should be reverted once
that is complete.

This patch passes the necessary configuration to devstack and tempest to
disable the advanced services from being started and also telling tempest
that the features aren't enabled. (which skips the tests) For setting the
enabled extensions in tempest this patch hard codes the extension list.
This is technically incorrect because this list isn't consistent between
branches, however since test-matrix.py doesn't support specifying
extensions yet, as a short-term temporary workaround it should be fine.

Related-Bug: #1400370
Co-Authored-By: Matthew Treinish <mtreinish@kortar.org>
Change-Id: I6044b27718de730286adcb69b2b1e1a2ab95e1fd