Handle more than one stream stored testrepository
This commit switches the cleanup function from having a hardcoded
assumption that there will only ever be 1 subunit stream stored in
the testrepository. Instead it uses the testr last command to retrieve
the most recent stream and uses that instead. This will be needed if
we start pre-seeding testrepository with data from previous runs.

Change-Id: I5a36e73998a9e9200df518c65ad43d1e6d3c13d4