remove heat from starting in grenade
It turns out there is actually no support for heat in grenade, so what
we're actually doing is starting heat on old, testing it, then not
actually upgrading heat, or testing it afterwards.

Testing heat on old isn't the purpose of this job, so we should drop
configuring that service. The normal heat tempest jobs should cover
that case anyway.

Change-Id: Iccc9135ef7abb73a414fa59ec7bc24513aa6e6e2