Allow PROJECTS to be overridden
Updating the default set of project repositories at the start of each
dsvm job adds 3-5m to the job's runtime.  For jobs that target only a
small subset of those repos, that overhead is largely unnecessary.  To
minimize the overhead of workspace setup, this change allows the
target project repositories (PROJECTS) to be overridden on a per-job
basis via the DEVSTACK_GATE_PROJECTS_OVERRIDE variable.

Change-Id: I0fb1bef66d82acdddeb9b29956fdd4311ea1f277