Add Neutron DVR aiopcpu support
Configure the subnode to run the extra services required
for Neutron DVR, like the L3 agent (in 'dvr' mode).

The data path between the controller and the subnode is
established via a tunnel created using OVS. The collective
troubleshooting effort [1] confirmed that n-net can work
successfully with gre tunnels being setup with ovs, therefore
only one logic is going to be maintained.

[1] https://review.openstack.org/#/c/171714/12

Depends-on: I6b4e5b82958e6d29dd450f1c4c9513f6a9e5053a
Change-Id: Id8d16877f09a1128c10f5c52438f6b7726c01f8e