Temporarily add trove plugin configuration
In order to move the trove configuration from devstack to trove,
we need to enable the plugin.  This should be done in a job config,
but since that can not be tested with a change and will break if
it lands before the plugin is moved, enable it in devstack-gate
to facilitate the transition.

Once the transition is complete, this can be removed.

Change-Id: I3506dec0e6097f9c2e9267110fdfb768faa23c85
Depends-On: I15667bf9c063e886755aef29056ef0028cbb1080