Add a loop to git_prune
git remote prune is sometimes hitting the same "GnuTLS recv error (-9):
A TLS packet with unexpected length was received" error that git remote
update was hitting in bug 1383403.

This change takes the timeout/retry logic in git_remote_update and puts
it into it's own git_timed method (sort of like the one in devstack)
so both git_remote_update and git_prune can leverage it.

Closes-Bug: #1455123

Change-Id: I026499c58baf5db644786b0cb4c6d04a5c70f7ae